Ensure flash object can access setFileFilters method before calling it.
For whatever reason, plupload triggers a Refresh event when handling a
StateChanged event. Strangely, a final StateChanged event is triggered
AFTER the UploadComplete event. If client code uses the UploadComplete
event to hide the flash object's parent container and the StateChanged event triggers a
Refresh using the flash runtime, then without checking for the presence
of #setFileFilters we get an exception unless the parent container is
visible.