Fix for HTML5 input z-index (issue #605)
To work properly in qTip's and jQuery modal dialogs in WebKit browsers
the input container must be 1 z-index below the button but not below
any other prior elements (like another qTip or a modal dialog's model
glasspane (the ui-widget-overlay).

This issue (#605) was caused because the existing code does not
consider the "current stacking context" when attempting to determine
the z-index of the browse button.

Here is a jsFiddle
(http://jsfiddle.net/stevemayhew/VQ5NP/4/embedded/result/) that fixes
the bug. With this verison of plupload and the snipit from the HTML5
change.  And lastly, here is the jsFiddle that demonstrates the bug:
http://jsfiddle.net/stevemayhew/PhCsp/embedded/result/