Automate Flash build
Linux is not the only one who deserves make build process a bit easier.
It is a Script that have a File Dialog to select Flex SDK folder and shows
friendly message error if compilation fails. (In most cases it fails because Java JDK is ausent or JAVA_HOME problems).