Complement to build.vbs
`mxmlc` is a too long line to be called from VBScript directly. The solution that I have found
was creating an external bat file that call the file.
VBScript do not read it response, but check if the file was created successfully.
If the file not exist, error message is thrown in VBScript.
The BAT is not closed automatically to allow user to read mxmlc output in prompt.