Enhanced support for hidden divs
Offset[Width/Height] and Client[Width/Height] can return 0 when it is
inside a hidden div that is show only as open (a "Modal" window, for
example). It is hidden, so Width and Height can't to be calculated. The
solution is try to show the div before call plupload
(https://github.com/moxiecode/plupload/issues/728) or set the Width and
Height manually on code and make plupload uses it (It is what this
solution purpose). If Width and Height is set on `browse_btn`, use it
instead of calculate Width and Height.