Add extensions to accepts attribute for html5.
Escape the madness of mime types by adding support for filtering by file extension to the html5 runtime!

This is part of the W3C spec, as documented here:
http://www.w3.org/TR/2012/WD-html5-20121025/states-of-the-type-attribute.html#attr-input-accept

At the moment the only browsers I know of that supports filtering by extension are Chrome and IE10. But hopefully support for other browsers should be in the pipeline.