UI: Bring consistent logic to max_file_count. Address #207, #342, #437, #838, #1002.
max_file_count will limit the number of files having status of QUEUED
or UPLOADING, that can be present in the queue at the same time.
Obviously after each upload cycle this number will auto-reset and it
will become possible to add more files. In case it is an undesirable
effect, max_file_count can be combined with multiple_queues which, when
set to false, limits upload cycles to just one.

Until now multiple_queues was set to false automatically, if
max_file_count was other than zero. Now we decoupled these two, so that
they could be used in an arbitrary combination to achieve a desired
result.