Triggering events in Plupload
Attempted to add a File Filter and use the `this.trigger()` causes an error. Even using `plupload.trigger()` caused an error because the method does not exist in plupload. It exists within `plupload.Uploader.prototype.__proto__.trigger`. This PR will fix that issue so that no one else goes as crazy as I did trying to find where the trigger method is at.

Unless of course we are using an older version of Plupload . . .