Fixed progress calculation when max_upload_slots > 1. This fixes the issue where the file progress would jump back and fourth as only completed chunks where included the calculation. It also fixes the total progress which would not include the last chunk.
The fix adds a loaded parameter to the chunk data structure. The loaded value is updated on each process event. When the progress is calculated the loaded value of each processing chunk is added with the size of the completed chunks.
A size parameter is also added to the chunk structure. This is used on completed chunks instead of calculating it on every event.